Move groovy spell tests into their own package
